# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API La Remise",
    "version": "1.0.0",
    "description": "API de gestion de la ressourcerie La Remise"
  },
  "paths": {
    "/objets": {
      "get": {
        "summary": "Liste tous les objets",
        "responses": {
          "200": { "description": "La liste des objets" }
        }
      },
      "post": {
        "summary": "Crée un nouvel objet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "libelle",
                  "poids_kg",
                  "etat_arrivee",
                  "categorie_id",
                  "depot_id"
                ],
                "properties": {
                  "libelle": { "type": "string" },
                  "poids_kg": { "type": "number" },
                  "etat_arrivee": {
                    "type": "string",
                    "enum": ["bon_etat", "a_reparer", "hors_service"]
                  },
                  "categorie_id": { "type": "integer" },
                  "depot_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Objet créé" },
          "400": { "description": "Champs manquants ou invalides" }
        }
      }
    },
    "/objets/{id}": {
      "get": {
        "summary": "Récupère un objet par son id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "L'objet correspondant" },
          "404": { "description": "Objet introuvable" }
        }
      },
      "put": {
        "summary": "Modifie un objet existant",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "libelle": { "type": "string" },
                  "poids_kg": { "type": "number" },
                  "etat_arrivee": {
                    "type": "string",
                    "enum": ["bon_etat", "a_reparer", "hors_service"]
                  },
                  "statut": {
                    "type": "string",
                    "enum": [
                      "arrive",
                      "en_reparation",
                      "en_rayon",
                      "vendu",
                      "recycle"
                    ]
                  },
                  "prix": { "type": "number" },
                  "date_mis_rayon": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Objet mis à jour" },
          "404": { "description": "Objet introuvable" }
        }
      },
      "delete": {
        "summary": "Supprime un objet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Objet supprimé" },
          "404": { "description": "Objet introuvable" }
        }
      }
    },
...
  }
}